move registration requests controller to api namespace
